adding refactor for column_major to row major index logic
